pyproject.toml
[project]
name = "sitplus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    #boilerplate
    "accelerate>=1.6.0",
    "bitsandbytes>=0.43.0",
    "dadaptation>=3.2",
    "diffusers>=0.33.1",
    #"flash-attn==2.7.2.post1",
    "ipython>=8.36.0",
    "lion-pytorch>=0.2.3",
    "lpips>=0.1.4",
    "lycoris-lora>=3.2.0.post2",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "prodigyopt>=1.1.2",
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "safetensors>=0.5.3",
    "setuptools>=80.4.0",
    "torch>=2.7.1", # nu torch for distributed ops tests
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "wandb>=0.19.11",
    #bespoke
    "gluon-experiment @ git+https://github.com/sqcu/gluon-experiment.git", # dion has rocked for a long long time
    #project related
    "pytorch-wavelets>=1.3.0",
    "PyWavelets>=1.6.0"
]

[project.optional-dependencies]
cuda = [
    "triton; sys_platform == 'linux'",
    "triton-windows<3.4; sys_platform == 'win32'",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
